LayoutTests/http/tests/dom/iframe-no-src-local-origin-allow-expected.txt

This tests that an iframe with no src attribute uses the parent document's origin when resolving links inserted into the frame via JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('theFrame').contentDocument.baseURI is "http://127.0.0.1:8000/dom/iframe-no-src-local-origin-allow.html"
PASS document.getElementById('theFrame').contentDocument.getElementById('innerFrame').contentDocument.body.innerHTML is 'PASS\n'
PASS successfullyParsed is true

TEST COMPLETE
